feat: implement useFallbackWallets hook for improved wallet detection
This update introduces the useFallbackWallets hook, which consolidates the fallback wallet definitions and enhances the detection logic in the ExternalWalletList component. The new implementation improves code organization and readability by centralizing wallet information retrieval, allowing for easier management of fallback wallets like Phantom, Coinbase, and OKX.

Changes:
- Added useFallbackWallets hook to manage fallback wallet data.
- Updated ExternalWalletList to utilize the new hook for rendering fallback wallets.
- Removed hardcoded fallback wallet definitions from ExternalWalletList.
