+++ packages/kit/src/components/WebDapp/index.ts
@@ -1,3 +1,4 @@
 export { ConnectWalletModal } from './ConnectWalletModal';
 export { ExternalWalletList } from './ExternalWalletList';
 export { OneKeyWalletConnectionOptions } from './OneKeyWalletConnectionOptions';
+export { useFallbackWallets } from './hooks/useFallbackWallets';
